Feature/api routes

+++ pages/api/recipes/[id].ts
@@ -1,6 +1,8 @@
 import dbConnect from '@/db/connect.ts';
 import Recipe from '@/db/models/Recipe';
 import { NextApiRequest, NextApiResponse } from 'next';
+import { getServerSession } from 'next-auth';
+import { authOptions } from '../auth/[...nextauth]';
 
 export default async function handler(
   request: NextApiRequest,
@@ -8,6 +10,7 @@ export default async function handler(
 ) {
   await dbConnect();
 
+  const session = await getServerSession(request, response, authOptions);
   const { id } = request.query;
 
   if (request.method === 'GET') {
@@ -16,9 +19,31 @@ export default async function handler(
       return response.status(404).json({ status: 'Not Found' });
     }
     response.status(200).json(recipe);
-  } else if (request.method === 'PUT') {
+  }
+
+  if (request.method === 'PUT') {
+    // Check if Session
+    if (!session) {
+      return response.status(401).json({ message: 'Unauthorized' });
+    }
+    // Check if author id is session user id and if fetch recipe
     try {
       const recipeData = request.body;
+      if (!recipeData.authorId || !session?.user?.id) {
+        return response.status(400).json({ message: 'Invalid Request' });
+      }
+
+      const existingRecipe = await Recipe.findById(id);
+
+      if (!existingRecipe) {
+        return response.status(404).json({ message: 'Recipe not found' });
+      }
+
+      if (recipeData.authorId !== session.user?.id) {
+        return response.status(403).json({ message: 'Forbidden' });
+      }
+
+      // Update Recipe
       await Recipe.findByIdAndUpdate(id, recipeData);
       return response.status(200).json({ status: `Recipe ${id} updated!` });
     } catch (error: unknown) {
@@ -28,8 +53,30 @@ export default async function handler(
         response.status(400).json({ error: 'an unknown error occured' });
       }
     }
-  } else if (request.method === 'DELETE') {
+  }
+
+  if (request.method === 'DELETE') {
+    // Check if Session
+    if (!session) {
+      return response.status(401).json({ message: 'Unauthorized' });
+    }
+    // Get the recipe by ID and check if the recipe Owner is equal to the Session User Id
     try {
+      const existingRecipe = await Recipe.findById(id);
+
+      if (!existingRecipe) {
+        return response.status(404).json({ message: 'Recipe not found' });
+      }
+
+      if (!existingRecipe.authorId || !session?.user?.id) {
+        return response.status(400).json({ message: 'Invalid Request' });
+      }
+      // Check if author id is session user id and if delete recipe
+      if (existingRecipe.authorId !== session.user?.id) {
+        return response.status(403).json({ message: 'Forbidden' });
+      }
+
+      // Delete Recipe
       await Recipe.findByIdAndDelete(id);
       return response
         .status(200)
